requirements.txt
beautifulsoup4==4.11.1
fasttext==0.9.2
nltk==3.6.5
numpy==1.21.2
python_Levenshtein==0.12.2
scipy==1.7.3
spacy==3.2.1